chore: integrate well with existing Vue apps
