+++ packages/api-reference/playground/multipage/src/Home.vue
@@ -1,0 +1,1 @@
+<template>Home</template>
