+++ packages/api-reference/playground/multipage/src/main.ts
@@ -2,24 +2,47 @@
 import '@scalar/components/style.css'
 import '@scalar/themes/style.css'
 import { createApp } from 'vue'
-import { createWebHashHistory } from 'vue-router'
+import { createRouter, createWebHistory } from 'vue-router'
 
-// import { createRouter, createWebHashHistory } from 'vue-router'
 import App from './App.vue'
+import Home from './Home.vue'
+import { ScalarApiReference, routesAsChildren } from './package'
 
-// import { routes } from './routes'
+/** Simulate a main app with an existing router, using just our component */
+const WITH_ROUTING = true
 
-const app = createApp(App, {
-  pages: 'single',
-  routePrefix: '/scalar',
-  history: createWebHashHistory(),
-})
+const app = createApp(
+  WITH_ROUTING ? App : ScalarApiReference,
+  WITH_ROUTING
+    ? {}
+    : {
+        pages: 'single',
+        history: createWebHistory(),
+      },
+)
 
-// const router = createRouter({
-//   history: createWebHashHistory(),
-//   routes: [],
-// })
+if (WITH_ROUTING) {
+  const router = createRouter({
+    history: createWebHistory(),
+    routes: [
+      {
+        path: '/',
+        name: 'home',
+        component: Home,
+      },
+      {
+        path: '/scalar',
+        name: 'scalar',
+        component: ScalarApiReference,
+        props: {
+          pages: 'multi',
+        },
+        children: routesAsChildren,
+      },
+    ],
+  })
 
-// app.use(router)
+  app.use(router)
+}
 
 app.mount('#app')
