+++ packages/api-reference/playground/multipage/src/package/index.ts
@@ -1,0 +1,2 @@
+export { default as ScalarApiReference } from './ScalarApiReference.vue'
+export { routesAsChildren } from './routes'
