+++ packages/api-reference/playground/multipage/src/package/views/Bar.vue
@@ -1,10 +1,10 @@
 <script lang="ts" setup>
-import Placeholder from '../components/Placeholder.vue'
+import Placeholder from '../../components/Placeholder.vue'
 </script>
 <template>
   <Placeholder
     class="m-1"
     style="height: 1200px">
-    Foo
+    Bar
   </Placeholder>
 </template>
